Use waitpid to poll for process liveness
This reverts commit 3c10cc0cb0a4738646966494fbeaab54bca115e1 which
move the process liveness checking implementation from JNR to JNA.
This broke on Windows as there is no shared libc available.

Instead, we move back to JNR, but using waitpid instead of getpgid to
perform the polling. This is a generic enough interfacec that it should
have very consistent support across POSIX operating systems and can even
be supported on Windows (in a future JNR release).

Finally, I believe the original motivation cited for moving to JNA in
3c10cc0 no longer applies; I have no trouble using this plugin with
current JNR and FreeBSD-11.0.

Fixes JENKINS-38682.